+++ JG/JohnnyGuitarNVSE.cpp
@@ -20,9 +20,9 @@
 #include "internal/decoding.h"
 #include "JohnnyGuitar/JohnnyEventPredefinitions.h"
 #include "JohnnyGuitar/misc.h"
+#include "JohnnyGuitar/EditorIDs.h"
 #include "JohnnyGuitar/JohnnyGuitarNVSE.h"
 #include "JohnnyGuitar/JohnnyParams.h"
-#include "JohnnyGuitar/EditorIDs.h"
 #include "JohnnyGuitar/JohnnyFunctions/fn_av.h"
 #include "JohnnyGuitar/JohnnyFunctions/fn_form.h"
 #include "JohnnyGuitar/JohnnyFunctions/fn_utility.h"
